hoot-rnd refactoring
