+++ hoot-rnd/src/main/cpp/hoot/rnd/cmd/LogoutCmd.cpp
@@ -30,7 +30,7 @@
 #include <hoot/core/cmd/BaseCommand.h>
 #include <hoot/core/util/Settings.h>
 #include <hoot/core/util/Log.h>
-#include <hoot/core/auth/HootServicesLoginManager.h>
+#include <hoot/rnd/auth/HootServicesLoginManager.h>
 #include <hoot/core/util/ConfigOptions.h>
 
 namespace hoot
